LayoutTests/discovery/discovery-zeroconf-servicefound.html

<html>
  <head>
    <script src="../resources/js-test-pre.js"></script>
    <script src="resources/discovery-testing.js"></script>
    <script type="text/javascript">
      description('Tests that servicefound event is send when a new service requested is published.');

      function onSuccess(services) {
          dumpServices(services);

          services.onservicefound = function() {
              debug("* service found");
              getNetworkServicesWithCallbacks(['zeroconf:_test-servicefound._tcp'], onSuccess2);
          }

          window.testRunner.addMockNetworkService('TestRunner3', '_test-servicefound._tcp', 'allowed', null);
      }

      function onSuccess2(services) {
          dumpServices(services);

          if (window.testRunner) {
              window.testRunner.notifyDone();
          }   
      }  

      function getServices() {
        getNetworkServicesWithCallbacks(['zeroconf:_test-servicefound._tcp'], onSuccess);
      }

      function buttonClicked() {

        window.testRunner.addMockNetworkService('TestRunner1', '_test-servicefound._tcp', 'allowed', null);
        window.testRunner.addMockNetworkService('TestRunner2', '_test-servicefound._tcp', 'allowed', null);

        // wait a while
        window.setTimeout(getServices, 2000);

        testRunner.waitUntilDone();
      }
    </script>
  </head>
<body onload="simulateButtonPress()">
<input type="button" id="button" value="Get Network Services" onmouseup="buttonClicked()">
<script src="../resources/js-test-post.js"></script>
</body>
<p id="description"></p>
</html>
